Inspections (#23)
* Convert fields to local

* Make final-able things final

* Redundant field initialisation

* size()==0 becomes isEmpty()

* Nullsafe comparison ordering

* Remove unnecessary super()

* Unnecessarily qualified static import

* Inverted if statement

* Simplify out if statements

* Make final

* Inverted negated if

* Inline one use variables

* Redundant throw

* Ensure exception root causes are passed on

* Formally ignore some ignored exceptions

* Public abstract constructors become protected

* Javadoc line breaks

* Enhanced switch statement

* Use pattern variables

* Use text blocks

* Use String.valueOf() rather than +""

* String.equals versus String.isEmpty

* Float cast to literal

* decapitalise

* Variable name fixes

* Inherit variable names hierarchically

* Variable name fixup

* Numeric cleanups

* replace construct/addAll with appropriate constructor

* String Builders

* Initialisations

* Inspections

* Make final

* Use isEmpty()

* Unnecessary toString()

* Redundant type casts

* Regex simplification

+++ src/net/coagulate/JSLBot/LLSD/LLSDBinary.java
@@ -50,7 +50,7 @@ public class LLSDBinary extends Atomic {
 		final byte[] by=toByte();
 		long longvalue=0;
 		for (int i=0;i<by.length;i++) {
-			longvalue+=((long) by[by.length-1-i]&0xffL)<<(8*i);
+			longvalue+=(by[by.length-1-i] &0xffL)<<(8*i);
 		}
 		return longvalue;
 	}
